server/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
web3==6.0.0
python-dotenv==1.0.0